pyproject.toml
[tool.poetry]
name = "practica-tesis"
version = "0.1.0"
description = ""
authors = ["LisandroLedesma <lichaledesma31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
scrapy = "^2.11.2"
selenium = "^4.24.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
